# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Machine à Café Connectée",
        "description": "API pour simuler une machine à café connectée",
        "contact": {
            "email": "contact@cafe-connecte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Serveur API"
        }
    ],
    "paths": {
...
        "/api/v1/coffees/{coffee}": {
            "get": {
                "tags": [
                    "Coffees"
                ],
                "summary": "Affiche les détails d'un café",
                "operationId": "82c9fd74f3d2bd3de31e7f80dd8313f6",
                "parameters": [
                    {
                        "name": "coffee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du café"
                    },
                    "404": {
                        "description": "Café non trouvé"
                    }
                }
            }
        },
        "/api/v1/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Liste toutes les commandes",
                "operationId": "17a756abe25d9d1b327f2f3f1f4ff8a1",
                "responses": {
                    "200": {
                        "description": "Liste des commandes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crée une nouvelle commande",
                "operationId": "98d74d33f0dfdd95d65841e55084c4da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "coffee_id"
                                ],
                                "properties": {
                                    "coffee_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "pickup_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-11 15:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commande créée"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/v1/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Affiche une commande spécifique",
                "operationId": "d95b0fa3593e90f7d275759583a60de8",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la commande"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Met à jour une commande",
                "operationId": "afffacb1c4f69451fb7eb58eff5db7d5",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coffee_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "customer_name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "pickup_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-05-11 15:30:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Commande mise à jour"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Supprime une commande",
                "operationId": "d418d7c2b44b7b122ac58d9fbf8084f1",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commande supprimée"
                    },
                    "404": {
                        "description": "Commande non trouvée"
                    },
                    "400": {
                        "description": "Impossible de supprimer une commande en cours"
                    }
                }
            }
        },
...
    },
...
}